dashboard/requirements.txt
streamlit==1.28.0
pandas==2.1.3
psycopg2-binary==2.9.9
plotly==5.18.0